Updated BackEnd Files along with Navbar
Changed Codes based on Pull Request.
