+++ Backend/DataAccessLayer/Models/LogoutRequest.cs
@@ -1,0 +1,9 @@
+﻿
+
+namespace DataAccessLayer.Models
+{
+    public class LogoutRequest
+    {
+        public string token;
+    }
+}
